@@ -1,7 +1,7 @@
 {
-  "name": "tinro",
-  "version": "0.6.12",
-  "description": "tinro is a tiny declarative router for Svelte",
+  "name": "tinro5",
+  "version": "0.0.1",
+  "description": "tinro5 is a tiny declarative router for Svelte 5",
   "main": "dist/tinro.js",
   "module": "dist/tinro.es.js",
   "svelte": "cmp/index.js",
@@ -15,9 +15,11 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/AlexxNB/tinro.git"
+    "url": "git+https://github.com/caracal7/tinro5.git"
   },
   "keywords": [
+    "tinro",
+    "tinro5",
     "svelte-router",
     "router",
     "svelte"
@@ -25,16 +27,16 @@
   "author": "Alexey Schebelev",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/AlexxNB/tinro/issues"
+    "url": "https://github.com/caracal7/tinro5/issues"
   },
-  "homepage": "https://github.com/AlexxNB/tinro#readme",
+  "homepage": "https://github.com/caracal7/tinro5#readme",
   "devDependencies": {
     "derver": "^0.5.3",
-    "esbuild": "^0.14.18",
-    "esbuild-svelte": "^0.6.2",
+    "esbuild": "^0.19.0",
+    "esbuild-svelte": "^0.9.1",
     "port-authority": "^1.1.2",
-    "puppeteer": "^13.1.3",
-    "svelte": "^3.46.4",
+    "puppeteer": "^21.0.0",
+    "svelte": "^5.23.0",
     "uvu": "^0.5.3"
   }
 }
