@@ -1,5 +1,7 @@
+
+import { mount } from 'svelte';
 import App from './App.svelte';
 
-new App({
-	target: document.body,
-});
+
+
+const app = mount(App, { target: document.body });
